requirements.txt
Django == 4.1.1
mysqlclient == 2.0.3
Pillow == 8.2.0
djangorestframework == 3.14.0
drf_yasg == 1.21.4
django-filter == 22.1
djangorestframework-simplejwt == 5.2.0
midtransclient == 1.1.1